Move rank calculation into new file
